requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.9.1
attrs==25.1.0
babel==2.17.0
bcrypt==4.3.0
bidict==0.23.1
black==25.1.0
blinker==1.9.0
boto3==1.37.9
botocore==1.37.9
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Deprecated==1.2.18
discord.py==2.4.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
eventlet==0.39.0
fastapi==0.115.8
filelock==3.17.0
Flask==2.2.2
Flask-Babel==2.0.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
fonttools==4.56.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
fuzzywuzzy==0.18.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-language==2.17.0
google-cloud-speech==2.31.1
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gspread==6.1.4
gTTS==2.2.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
Levenshtein==0.26.1
line-bot-sdk==2.2.1
lxml==5.3.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mecab-python3==1.0.10
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.2
python-jose==3.4.0
python-Levenshtein==0.26.1
python-multipart==0.0.20
python-socketio==5.12.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.28.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.45.3
stripe==5.1.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==2.2.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
